GSYE-258: Refactored classes that generate plotly graphs from the export module to their own module.
